Fix reporting latency to plugin host

+++ src/processors/chain/DryWetProcessor.cpp
@@ -17,12 +17,12 @@ void DryWetProcessor::copyDryBuffer (const AudioBuffer<float>& buffer)
     dryBuffer.makeCopyOf (buffer, true);
 }
 
-void DryWetProcessor::processBlock (AudioBuffer<float>& buffer, float latencySamples)
+void DryWetProcessor::processBlock (AudioBuffer<float>& buffer, int latencySamples)
 {
     const auto numChannels = buffer.getNumChannels();
     const auto numSamples = buffer.getNumSamples();
 
-    dryDelay.setDelay (latencySamples);
+    dryDelay.setDelay ((float) latencySamples);
     dsp::AudioBlock<float> dryBlock (dryBuffer);
     dsp::ProcessContextReplacing<float> dryContext (dryBlock);
     dryDelay.process (dryContext);
